Updated documentation for rmqsBuffer_t

+++ src/rawClient/rmqsBuffer.h
@@ -22,6 +22,12 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 ****************************************************************************/
 //---------------------------------------------------------------------------
+/**
+ * @file rmqsBuffer.h
+ *
+ * Memory streams handling
+ */
+//---------------------------------------------------------------------------
 #ifndef rmqsBufferH
 #define rmqsBufferH
 //---------------------------------------------------------------------------
@@ -30,10 +36,11 @@ SOFTWARE.
 //---------------------------------------------------------------------------
 #include "rmqsGlobal.h"
 //---------------------------------------------------------------------------
-/** \brief Holds a memory stream in which information can be written in append mode and that
- *  dynamically allocates the memory necessary to contain it. Memory is allocated in chunks of customizable size
- *  to reduce the number of allocations needed
+/** \brief A memory I/O stream
  *
+ * Holds a memory stream in which information can be written in append mode and that
+ * dynamically allocates the memory necessary to contain it. Memory is allocated in chunks of customizable size
+ * to reduce the number of allocations needed
  * \param ChunkSize The initial amount of memory that is allocated when the structure is created or when new memory blocks need
  * to be allocated
  * \param Size The size of the data written to the buffer
@@ -49,10 +56,10 @@ SOFTWARE.
  */
 typedef struct
 {
-    uint32_t ChunkSize;
-    uint32_t Size;
-    uint32_t Position;
-    uint32_t CurrentMemorySize;
+    size_t ChunkSize;
+    size_t Size;
+    size_t Position;
+    size_t CurrentMemorySize;
     void *Data;
     uint32_t Tag1;
     uint32_t Tag2;
@@ -60,14 +67,87 @@ typedef struct
 }
 rmqsBuffer_t;
 //---------------------------------------------------------------------------
+/** \brief Create a new memory stream
+ *
+ * \return Pointer to the newly created memory stream
+ *
+ */
 rmqsBuffer_t * rmqsBufferCreate(void);
+//---------------------------------------------------------------------------
+/** \brief Destroys a previously created memory stream
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream to delete
+ * \return void
+ *
+ */
 void rmqsBufferDestroy(rmqsBuffer_t *Buffer);
+//---------------------------------------------------------------------------
+/** \brief Clears the content of a memory stream
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream to clear
+ * \param ResetMemoryBuffer bool_t If the allocated memory needs to be freed, set false to avoid a new allocation of memory on write
+ * \return void
+ *
+ */
 void rmqsBufferClear(rmqsBuffer_t *Buffer, const bool_t ResetMemoryBuffer);
+//---------------------------------------------------------------------------
+/** \brief Set the three 'Tag' variables of the stream to zero
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream
+ * \return void
+ *
+ */
 void rmqsBufferClearTags(rmqsBuffer_t *Buffer);
+//---------------------------------------------------------------------------
+/** \brief Reads a certain amount of bytes from the stream, starting at
+ *         the current location
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream to read from
+ * \param ReadBuffer void* Memory block in which the bytes must be copied
+ * \param NoOfBytes size_t How many bytes must be read from the stream
+ * \return The amount of bytes read from the stream
+ *
+ */
 size_t rmqsBufferRead(rmqsBuffer_t *Buffer, void *ReadBuffer, size_t NoOfBytes);
+//---------------------------------------------------------------------------
+/** \brief Write a certain amount of bytes to the stream, starting from
+ *         the current location
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream to write to
+ * \param WriteBuffer void* Memory block from which the bytes are to be copied
+ * \param NoOfBytes size_t How many bytes must be written to the stream
+ * \return The amount of bytes written to the stream
+ *
+ */
 size_t rmqsBufferWrite(rmqsBuffer_t *Buffer, void *WriteBuffer, size_t NoOfBytes);
+//---------------------------------------------------------------------------
+/** \brief Deletes a certain amount of bytes from the stream, starting from
+ *         the beginning of the stream
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream
+ * \param NoOfBytes size_t How many bytes must be deleted from the stream
+ * \return void
+ *
+ */
 void rmqsBufferDelete(rmqsBuffer_t *Buffer, size_t NoOfBytes);
+//---------------------------------------------------------------------------
+/** \brief Set the current position of the memory stream
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream
+ * \param Position size_t New position within the memory stream
+ * \return void
+ *
+ */
 void rmqsBufferMoveTo(rmqsBuffer_t *Buffer, size_t Position);
+//---------------------------------------------------------------------------
+/** \brief Set the size of the allocated memory of the stream.
+ *  This function should be called only from the other stream functions.
+ *
+ * \param Buffer rmqsBuffer_t* Pointer to the memory stream
+ * \param RequiredMemorySize size_t Size of the allocated memory
+ * \return void
+ *
+ */
 void rmqsBufferSetMemorySize(rmqsBuffer_t *Buffer, const size_t RequiredMemorySize);
 //---------------------------------------------------------------------------
 #endif
